pyproject.toml
[project]
name = "dbrownell-toolsdirectory"
version = "0.0.0"
#          ^^^^^
# Wheel names will be generated according to this value. Do not manually modify this value; instead
# update it according to committed changes by running this command from the root of the repository:
#
#   uv run python -m AutoGitSemVer.scripts.UpdatePythonVersion ./pyproject.toml ./src

description = "Creates functionality to manipulate the shell in support of tools defined within a tools directory."
readme = "README.md"
authors = [
    { name = "David Brownell", email = "github@DavidBrownell.com" }
]
requires-python = ">= 3.14"
dependencies = [
    "python-dotenv>=1.2.1",
    "semantic-version>=2.10.0",
    "typer>=0.20.1",
]
classifiers = [
    "Operating System :: MacOS",
    "Operating System :: Microsoft :: Windows",
    "Operating System :: POSIX :: Linux",
    "Programming Language :: Python",
    "Programming Language :: Python :: 3.14",
]

[project.license]
text = "MIT"

[project.urls]
Homepage = "https://github.com/davidbrownell/dbrownell_ToolsDirectory"
Documentation = "https://github.com/davidbrownell/dbrownell_ToolsDirectory"
Repository = "https://github.com/davidbrownell/dbrownell_ToolsDirectory"

[build-system]
requires = ["uv_build>=0.9.18,<0.10.0"]
build-backend = "uv_build"

[dependency-groups]
dev = [
    "autogitsemver>=0.9.2",
    "dbrownell-commitemojis>=0.2.0",
    "pre-commit>=4.5.1",
    "pyfakefs>=6.0.0",
    "pytest>=9.0.2",
    "pytest-cov>=7.0.0",
    "ruff>=0.14.10",
    "ty>=0.0.8",
]

[tool.pytest.ini_options]
addopts = "--verbose -vv --capture=no --cov=dbrownell_ToolsDirectory --cov-report html --cov-report term --cov-report xml:coverage.xml --cov-fail-under=95.0"
python_files = [
    "**/*Test.py",
]

[tool.ruff]
line-length = 110

[tool.ruff.lint]
exclude = ["tests/**"]

select = ["ALL"]

ignore = [
    "ANN002", # Missing type annotation for `*args`
    "ANN003", # Missing type annotation for `**kwargs`
    "BLE001", # Do not catch blind exception: `Exception`
    "COM812", # Trailing comma missing
    "D104", # Missing docstring in public package
    "D105", # Missing docstring in magic method
    "D107", # Missing docstring in `__init__` method
    "D202", # No blank lines allowed after function docstring
    "E501", # Line too long
    "FIX002", # Line contains TODO, consider resolving the issue
    "I001", # Import block is un-sorted or un-formatted
    "N802", # Function name `xxx` should be lowercase
    "N999", # Invalid module name
    "RSE102", # Unnecessary parentheses on raise exception
    "S101", # Use of assert detected
    "TC006", # Add quotes to type expression in `typing.cast()`
    "TD002", # Missing author in TODO
    "TD003", # Missing issue link for this TODO
    "TRY002", # Create your own exception
    "TRY300", # Consider moving this statement to an `else` block
    "UP032", # Use f-string instead of `format` call
]

[tool.ruff.lint.mccabe]
max-complexity = 15

[tool.ruff.lint.pylint]
max-args = 10
max-branches = 20
max-returns = 20

[tool.uv.build-backend]
module-name = "dbrownell_ToolsDirectory"
